ROL: Fixed signed/unsigned comparison warning in LineSearchStep.  I also added an objective evaluation in the update functions for the new unglobalized steps. This will be useful for the case when the objective function is changed between every iteration.  This may result in an additional objective function evaluation when using line search with an objective function that does not change.

+++ packages/rol/src/step/ROL_LineSearchStep.hpp
@@ -461,7 +461,7 @@ public:
     std::string desc = desc_->print(algo_state,false);
     desc.erase(std::remove(desc.end()-3,desc.end(),'\n'), desc.end());
     std::string name = desc_->printName();
-    int pos = desc.find(name);
+    size_t pos = desc.find(name);
     if ( pos != std::string::npos ) {
       desc.erase(pos, name.length());
     }
